requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
